# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Vendor Service API
  description: |
    Microservicio de gestión de proveedores.
    Proporciona CRUD de proveedores, sucursales, comisiones, categorías,
    políticas, métodos de pago y carga de logotipos.
  version: "1.0.0"
  contact:
    name: Vendor Service Team
servers:
  - url: http://localhost:3001
    description: Desarrollo local
  - url: https://api.example.com
    description: Producción
paths:
...
  /api/vendors:
    post:
      tags: [Vendors]
      summary: Crear un nuevo proveedor
      operationId: createVendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorRequest"
            example:
              vendor_name: "Distribuidora ABC"
              vendor_email: "contacto@abc.com"
              vendor_ruc: "12345678901"
              user_id: 42
              vendor_phone: "+525512345678"
              vendor_address: "Av. Reforma 123, CDMX"
              categories: [1, 3]
      responses:
        "201":
          description: Proveedor creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"

    get:
      tags: [Vendors]
      summary: Listar proveedores
      operationId: listVendors
      parameters:
        - name: vendor_status
          in: query
          schema:
            type: string
            enum: [PENDING, ACTIVE, INACTIVE, SUSPENDED]
          description: Filtrar por estado
        - name: vendor_email
          in: query
          schema:
            type: string
            format: email
          description: Filtrar por email
        - name: vendor_ruc
          in: query
          schema:
            type: string
          description: Filtrar por RUC
      responses:
        "200":
          description: Lista de proveedores
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vendor"
...
  /api/vendors/{vendor_id}/categories:
    get:
      tags: [Categories]
      summary: Obtener categorías de un proveedor
      operationId: getVendorCategories
      parameters:
        - name: vendor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Categorías asignadas al proveedor
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        categories:
                          $ref: "#/components/schemas/Category"

    put:
      tags: [Categories]
      summary: Reemplazar categorías de un proveedor
      operationId: replaceVendorCategories
      parameters:
        - name: vendor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_ids]
              properties:
                category_ids:
                  type: array
                  items:
                    type: integer
                  description: IDs de categorías a asignar
            example:
              category_ids: [1, 2, 5]
      responses:
        "200":
          description: Categorías actualizadas correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Categorías actualizadas correctamente"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ──────────────────────────────────────────────
  # POLICIES
  # ──────────────────────────────────────────────
...
  /api/vendors/{vendor_id}/payment-methods:
    get:
      tags: [Payment Methods]
      summary: Obtener métodos de pago de un proveedor
      operationId: getVendorPaymentMethods
      parameters:
        - name: vendor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Métodos de pago del proveedor
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PaymentMethod"

    put:
      tags: [Payment Methods]
      summary: Reemplazar métodos de pago de un proveedor
      operationId: replaceVendorPaymentMethods
      parameters:
        - name: vendor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_method_ids]
              properties:
                payment_method_ids:
                  type: array
                  items:
                    type: integer
                  description: IDs de métodos de pago a asignar
            example:
              payment_method_ids: [1, 3, 4]
      responses:
        "200":
          description: Métodos de pago actualizados
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Métodos de pago actualizados correctamente"
        "400":
          $ref: "#/components/responses/ValidationError"

  # ──────────────────────────────────────────────
  # LOGO
  # ──────────────────────────────────────────────
...
components:
  schemas:
    Vendor:
      type: object
      properties:
        vendor_id:
          type: integer
          example: 1
        vendor_name:
          type: string
          example: "Distribuidora ABC"
        vendor_ruc:
          type: string
          example: "12345678901"
        vendor_email:
          type: string
          format: email
          example: "contacto@abc.com"
        vendor_phone:
          type: string
          example: "+525512345678"
        vendor_address:
          type: string
          example: "Av. Reforma 123, CDMX"
        vendor_status:
          type: string
          enum: [PENDING, ACTIVE, INACTIVE, SUSPENDED]
          example: "ACTIVE"
        vendor_logo_url:
          type: string
          format: uri
          nullable: true
        user_id:
          type: integer
          example: 42
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateVendorRequest:
      type: object
      required: [vendor_name, vendor_email, vendor_ruc, user_id]
      properties:
        vendor_name:
          type: string
          description: Nombre del proveedor
        vendor_email:
          type: string
          format: email
          description: Correo electrónico del proveedor
        vendor_ruc:
          type: string
          description: RUC del proveedor
        user_id:
          type: integer
          description: ID del usuario asociado
        vendor_phone:
          type: string
          description: Teléfono del proveedor
        vendor_address:
          type: string
          description: Dirección del proveedor
        categories:
          type: array
          items:
            type: integer
          description: IDs de categorías a asignar
...
    VendorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: "#/components/schemas/Vendor"
...
    Category:
      type: object
      properties:
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: "Electrónicos"
        description:
          type: string
          example: "Productos electrónicos y accesorios"
...
    PaymentMethod:
      type: object
      properties:
        payment_method_id:
          type: integer
          example: 1
        payment_method_name:
          type: string
          example: "Efectivo"
        code:
          type: string
          example: "CASH"
        description:
          type: string
        is_active:
          type: boolean
          example: true
  responses:
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "vendor_name es requerido; Email válido es requerido; RUC es requerido"
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Proveedor no encontrado"
    ConflictError:
      description: Conflicto (recurso duplicado)
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "El email ya está registrado por otro proveedor"
...